+++ ros2/protobuf/BUILD.bazel
@@ -1,0 +1,35 @@
+load("@rules_python//python:defs.bzl", "py_binary", "py_library")
+load("@rules_ros2_pip_deps//:requirements.bzl", "requirement")
+
+exports_files([
+    "protobuf.bzl",
+])
+
+py_library(
+    name = "proto_to_ros2",
+    srcs = ["proto_to_ros2.py"],
+    deps = [
+        "@com_google_protobuf//:protobuf_python",
+    ],
+)
+
+py_binary(
+    name = "proto_to_ros2_msg",
+    srcs = ["proto_to_ros2_msg.py"],
+    visibility = ["//visibility:public"],
+    deps = [
+        ":proto_to_ros2",
+        "@com_google_protobuf//:protobuf_python",
+    ],
+)
+
+py_binary(
+    name = "proto_to_ros2_converter",
+    srcs = ["proto_to_ros2_converter.py"],
+    visibility = ["//visibility:public"],
+    deps = [
+        ":proto_to_ros2",
+        "@com_google_protobuf//:protobuf_python",
+        requirement("empy"),
+    ],
+)
